Working raycast

+++ dev/include/architecture/bbox.h
@@ -9,12 +9,14 @@ typedef struct bbox
 {//either children or tris is null
     float min[3];
     float max[3];
-    size_t total;
-    size_t maxtotal;
+    //total number of triangles in the bbox (including children)
+    int total;
+    //max number of triangles in the bbox (including children)
+    int maxtotal;
     struct bbox ** children;
     triangle * tris;
-    //also maxtotal
-    size_t c_size;
+    //number of children (either bbox or tris) in the bbox
+    int c_size;
 }bbox;
 void free_bbox(bbox *b);
